RUI-000: rename kpi and export it
